requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-CORS==4.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
markupsafe==2.0.1
SQLAlchemy==1.4.50